+++ src/frontend/apps/e2e/__tests__/app-impress/doc-header.spec.ts
@@ -395,7 +395,9 @@ test.describe('Doc Header', () => {
       navigator.clipboard.readText(),
     );
     const clipboardContent = await handle.jsonValue();
-    expect(clipboardContent.trim()).toBe(`<h1>Hello World</h1><p></p>`);
+    expect(clipboardContent.trim()).toBe(
+      `<h1 data-level=\"1\">Hello World</h1><p></p>`,
+    );
   });
 
   test('it checks the copy link button', async ({ page }) => {